// src/app/interface/customer-interface.ts
export interface Customer{
    CustomerId: number;
    CustomerCompetencyId: number;
    customerCompetencyType: string;
    CustomerName: string;
    CustomerSurname: string;
    CustomerPhoneNo: string;
    CustomerEmail: string;
    CustomerIdNumber: string;
    AddressId: number;
    streetName: string;
    streetNumber: string;
    TitleId: number;
    titleDescr: string;
}